feat(printer-settings): enhance PrinterSettings interface and update related components
- Added new typography options (fontSize, fontFamily, lineHeight, enableBold) to PrinterSettings.
- Updated PrinterSettingsTab to include a save settings handler.
- Modified BasicPrintSettingsCard to support saving settings.
- Refactored receipt generation utilities to separate data and HTML generation.
- Removed deprecated receipt-generator file.
